Separated main into user functions and image processing functions to ease reading
